requirements.txt
# ==============================================================================
# UNIFIED REQUIREMENTS FILE - WORKS LOCALLY AND ON RENDER
# ==============================================================================

# ==============================================================================
# WEB FRAMEWORK & API
# ==============================================================================
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
aiofiles==24.1.0
starlette>=0.47.0,<1.0.0

# ==============================================================================
# MACHINE LEARNING & AI
# ==============================================================================
# TensorFlow - CPU version for cloud deployment compatibility
# Works both locally and on cloud platforms like Render
tensorflow-cpu>=2.16.0,<2.18.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<1.8.0

# ==============================================================================
# IMAGE PROCESSING
# ==============================================================================
pillow>=10.0.0,<12.0.0
split-folders==0.5.1

# ==============================================================================
# DATA PROCESSING
# ==============================================================================
pandas>=2.0.0,<3.0.0

# ==============================================================================
# DATABASE & BACKEND SERVICES
# ==============================================================================
# Supabase client and dependencies
supabase==2.17.0
storage3==0.12.0
postgrest==1.1.1
gotrue==2.12.3
realtime==2.6.0
supafunc==0.10.1

# ==============================================================================
# HTTP & NETWORKING
# ==============================================================================
requests>=2.30.0,<3.0.0
httpx>=0.25.0,<1.0.0

# ==============================================================================
# DATA VALIDATION & SERIALIZATION
# ==============================================================================
pydantic>=2.0.0,<3.0.0

# ==============================================================================
# UTILITIES
# ==============================================================================
python-dateutil>=2.8.0,<3.0.0

# ==============================================================================
# DEPLOYMENT NOTES:
# ==============================================================================
# This file uses tensorflow-cpu which works on:
# ✓ Local development (Windows/Mac/Linux)
# ✓ Cloud platforms (Render, Heroku, Railway, etc.)
# ✓ Docker containers
# ✓ GitHub Actions/CI
#
# Environment variables for deployment:
# - SUPABASE_URL=your_supabase_url
# - SUPABASE_KEY=your_supabase_key
# - TF_CPP_MIN_LOG_LEVEL=2
# - PYTHONUNBUFFERED=1
